# Avoid indirection for value hashes