--- a/gen/DSE.Open.Values.Generators/ValueTypesGenerator.Emitter.cs
+++ b/gen/DSE.Open.Values.Generators/ValueTypesGenerator.Emitter.cs
@@ -226,7 +226,7 @@ public partial class ValueTypesGenerator
                 }
 
                 writer.WriteBlock("""
-                                      return HashCode.Combine(_value);
+                                      return _value.GetHashCode();
                                   }
                                   """);
             }